Use 32bit integer for index types in DQM rather than 64bit

+++ dimod/discrete/cydiscrete_quadratic_model.pxd
@@ -24,7 +24,7 @@ from dimod.bqm.cppbqm cimport AdjVectorBQM as cppAdjVectorBQM
 from dimod.bqm.common cimport Integral32plus, Numeric, Numeric32plus
 
 ctypedef np.float64_t bias_type
-ctypedef np.int64_t index_type  # todo: int32
+ctypedef np.int32_t index_type  # todo: int32
 
 ctypedef fused Unsigned:
     np.uint8_t
